fix: return jsonify

+++ ai/face_generate.py
@@ -1,4 +1,4 @@
-from flask import Flask, request, send_file, make_response
+from flask import Flask, request, jsonify
 import json
 import sys
 import os
@@ -6,7 +6,8 @@ sys.path.append('./face_maker/DualStyleGAN')
 from PIL import Image
 import base64
 from face_maker.DualStyleGAN.style_transfer import *
-import io
+import io, glob
+import requests
 
 app = Flask(__name__)
 
@@ -18,31 +19,38 @@ def home():
 @app.route('/generate_image', methods = ['POST']) #flask는 단일스레드+동기처리를 수행하기 때문에 face_maker는 항상 하나의 요청에서만 접근된다.
 def generate_image():
 
+    print('----------------------------------------\n')
     print(type(request.files['image']))
     print(request.values.keys())
-    file_path = request.values['user_id']+'.jpg'
-    request.files['image'].save(file_path)
-    virtual_face = face_maker.generate(file_path,int(request.values['style_id']))
+    print('----------------------------------------\n')
 
+    # 파일 저장 위치 만들기
+    file_path = request.values['user_id'] + '.jpeg'
+
+    # 파일 저장
+    request.files['image'].save(file_path)
+        
+    # 관상 이미지 생성
+    virtual_face = face_maker.generate(file_path, int(request.values['style_id']))
+    
+    # Image 인스턴스로 변환
     pil_virutal_face = Image.fromarray(virtual_face.astype('uint8'))
     file_object = io.BytesIO()
 
-    # write JPEG in file-object
-    pil_virutal_face.save(file_object, 'JPEG')
-    
-    # move to beginning of file so `send_file()` it will read from start
-    file_object.seek(0)
+    # 이미지를 바이너리로 저장
+    pil_virutal_face.save(file_object, format='JPEG')
+    image_binary = file_object.getvalue()
 
-    response = make_response(file_object)
+    response_data = {'image_binary':base64.b64encode(image_binary).decode('utf-8')}
 
-    response.headers.set('Content-Type', 'image/jpeg')
-    response.headers.set(
-        'Content-Disposition', 'attachment', filename='generated_image')
+        # 기존 사진 삭제
+    if os.path.exists(file_path):
+        os.remove(file_path)
+    else:
+        pass
     
-    #remove temp jpg file
-    os.remove(file_path)
+    return jsonify(response_data)
     
-    return response #send_file(file_object, mimetype='image/jpeg')
 
 if __name__ == '__main__':
     face_maker = StyleTransfer()
